pyproject.toml
[tool.poetry]
name = "memory"
version = "0.1.0"
description = ""
authors = ["Gianluca Rossi <gr.gianlucarossi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.310"
pypdf = "^3.16.2"
sentence-transformers = "^2.2.2"
huggingface-hub = "^0.17.3"
python-dotenv = "^1.0.0"
chromadb = "^0.4.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
